+++ src/rooms.c
@@ -5,13 +5,13 @@
 ** Ant-based calculation unit
 */
 
-#include "mysh.h"
+#include "lemin.h"
 
 char *getroom_tunnel(char *input)
 {
     char *room = NULL;
-    unsigned int end_room = 0;
-    unsigned int len = my_strlen(input);
+    size_t end_room = 0;
+    size_t len = my_strlen(input);
 
     if (input[0] == ' ' || input[0] == '-' || input[0] == '\n' || !len)
         return NULL;
@@ -24,14 +24,13 @@ char *getroom_tunnel(char *input)
     return room;
 }
 
-char *getroom2(unsigned int end_room, char *input,
-unsigned int num_words, bool new)
+char *getroom2(size_t end_room, char *input, size_t num_words, bool new)
 {
-    unsigned int len = my_strlen(input);
+    size_t len = my_strlen(input);
     char *room = malloc(sizeof(char) * (end_room + 1));
     my_strncpy(room, input, end_room);
     room[end_room] = '\0';
-    for (unsigned int i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
         if (input[i] == ' ' || input[i] == '-' || input[i] == '\n') {
             new = true;
         } if (new && input[i] != ' ' && input[i] != '-' && input[i] != '\n') {
@@ -51,9 +50,9 @@ unsigned int num_words, bool new)
 
 char *getroom(char *input)
 {
-    unsigned int end_room = 0;
-    unsigned int len = my_strlen(input);
-    unsigned int num_words = 0;
+    size_t end_room = 0;
+    size_t len = my_strlen(input);
+    size_t num_words = 0;
     bool new_word = true;
 
     if (input[0] == ' ' || input[0] == '-' || input[0] == '\n' || !len)
@@ -77,14 +76,14 @@ int create_rooms(var_t *var, char *data)
         write(2, "Error: Invalid room.\n", 21);
         return 84;
     }
-    for (unsigned int i = 0; var->room && var->room[i]; i++) {
+    for (size_t i = 0; var->room && var->room[i]; i++) {
         if (!my_strcmp(var->room[i]->data, data)) {
             write(2, "Error: Room already exists.\n", 28);
             return 84;
         }
     }
     new_rooms = malloc(sizeof(link_t *) * (unsigned long)(var->room_nb + 2));
-    for (unsigned int i = 0; i < var->room_nb; i++)
+    for (size_t i = 0; i < var->room_nb; i++)
         new_rooms[i] = var->room[i];
     new_rooms[var->room_nb] = create_link(data);
     var->room_nb++;
